nrt: reserve: add cache implementation
Implement numa-aware pessimistic overallocation cache,
using the reserve plugin hook points.

See `pkg/noderesourcetopology/docs/` for the full design
and rationale of this cache.

To resync the cache with the actual node state, we need
a periodic resync loop; the period is controlled with
a user-exposed parameter which, alongside the reserve plugin,
control the functionality.

In other words, to enable this caching, the user needs to BOTH
- enable the noderesourcetopology reserve plugin
- configure the resync period using the new parameter

failing to do *both* will make the scheduler misbehave, but
it seems there is not a convenient way to enforce this requirement.
We will try to improve this UX issue in the future.

Signed-off-by: Francesco Romani <fromani@redhat.com>
(cherry picked from commit 0f4ed7a9713d0cf984fa43dd16c4adb8ac05ef14)