pfp dump support
Leverage podfingerprint >= v0.2.2 to enable async dump of computed
podfingerprint status to make it easier to understand and
troubleshoot the cache management and cache desync.

In addition to the regular logging, there is the option now
to enable this extra debug using an environment variable.
The variable is used to avoid adding new binary flags,
maximizing the backward compatibility.

We expect the status dump directory to be a fixed-size memory
bound filesystem, to avoid clogging the filesystem and to
automatically dispose leftovers.

Note the dump is best-effort: failed writes are silently ignored
Note the dump contains a timestamp taken at write time, not at
compute time

Signed-off-by: Francesco Romani <fromani@redhat.com>