nrt: overreserve: refine foreign pods detection
"foreign pods" are pods handled by other schedulers, namely
the primary default scheduler.
"foreign pods" need to be tracked to make sure the overreserve
cache is accurate and trusted. When a foreign pod is detected,
the relevant node cache is marked untrusty and verified periodically
using the standard resync method.

Expose in the configuration the foreign pods setting.
Previously, we didn't had this option in the configuration
because the foreign pod detection was simple, thus always
enabled when the cache was in turn enabled.

Tuning the foreign pods detection (to reduce the scheduling latency)
benefits now from configurability.

The default is the existing behavior, and the cluster admin
can now disable entirely the foreign pods detection (perhaps because
the NRT plugin is consumed by the main/only scheduler, thus is
redundant) or enable it only for pods which have exclusive resources,
which is indeed a way to greatly reduce the resync churn and thus
the scheduling latency.

Finally, the last addition is the option to only consider pods
which have exclusive resources assigned to them. This is because
only pods which have exclusive resources assigned to them can have
NUMA-affine resources, thus contribute to the overreserve accounting.

If we ignore pods with non-exclusive resources, we can significantly
reduce noise, churn and CPU time.

Signed-off-by: Francesco Romani <fromani@redhat.com>
(cherry picked from commit abb5beca78ea46947acb54a60f565cc162040c3f)