[KNI][downstream] nrt: add optional kni-specific informer
The podinformer supplied by the scheduler framework filters out
pods in terminal state. Changing this is nontrivial, and we need
a fix on older versions.

In order to completely fix the non-running pod desync, we need
to either get pods in terminal phase on the scheduler side or
to filter out these pods on nodes.

The best option here is adding a separate independent dedicated
informer on the scheduler side. This is the more self contained
change AND the one which scales better, requiring "just" one
additional subscription to the apiserver (vs N=len(workers)).

The extra dedicated informer can be enabled using the environment
variable `NRT_ENABLE_INFORMER=<bool>`.
If disabled, the code behaves exactly like previously,
using the scheduler-framework provided informer.

We still need to pursue options in the context of the scheduler
framework to see which improvements we can land in there.

Signed-off-by: Francesco Romani <fromani@redhat.com>