nrt: cache: discardreserved: fix return value
The caches are authoritative regarding the freshness of
the data they return. This is the intended meaning of the
boolean return value in the GetCachedNRTCopy signature, kinda
like "isValid" or "isFresh". So, when there's no data in
the cache, the information *IS* fresh and should be reported as such.

The cached data is expected to be out of data ("stale", "not fresh",
"invalid") only when foreign pods (pods running on a node managed
by the scheduler, but not scheduled by this scheduler) are detected,
and even so, some cache implementations may tolerate this condition
and still report valid data.

Signed-off-by: Francesco Romani <fromani@redhat.com>
(cherry picked from commit aa79c7f091f50ffdd4ae39ac5ab6c7f2ab5a016f)