[KNI] nrtpatch.sh: read from stdin only once
stdin is a stream so reading it multiple times
from it - even using the `/dev/stdin` pseudo files
won't work. The second read and following will
just yield EOF, making the script fail.

We need to read it once, buffer everything
and reuse that buffer.

Considering the nature of this script and the expected
datasize involved, this seems still reasonnable.

Signed-off-by: Francesco Romani <fromani@redhat.com>